requirements.txt
Django==1.9.4
djangorestframework==3.3.3
fn.py==0.4.4
psycopg2==2.6.1
xlrd==0.9.4
